feat: Add user menu and mobile menu functionality to the Navbar component
This commit adds functionality to the Navbar component to handle user menu and mobile menu. It includes the necessary state variables and event handlers to toggle the visibility of the user menu and mobile menu. The user menu is triggered by clicking a button, and it displays a list of options for the user. The mobile menu is triggered by clicking a button and it expands to show a list of navigation links. The visibility of both menus is controlled by the respective state variables.