Refactor wishlistSlice to handle book wishlist functionality
In this commit, I have refactored the wishlistSlice to handle the book wishlist functionality. The wishlistSlice now includes the necessary reducers for adding and removing book IDs from the wishlist. The state structure has been updated to store an array of book IDs in the wishlist. I have also added a selector function to retrieve the wishlist from the RootState.

- Added addToWishlist reducer to add a book ID to the wishlist
- Added removeFromWishlist reducer to remove a book ID from the wishlist
- Updated the state structure to store an array of book IDs
- Created selectWishlist selector to retrieve the wishlist from the RootState

This commit prepares the wishlistSlice for integrating with the UI components that interact with the wishlist feature.